updated, tested without errors

+++ RRSVM/Tests/MaxPoolingTest.py
@@ -13,6 +13,6 @@ if __name__ == '__main__':
     n_channel = 100
     feature_size = 6
     batch_size = 3
-    input = (Variable(torch.FloatTensor(torch.randn(1, n_channel, feature_size, feature_size)), requires_grad=True), kernel_size)
+    input = (Variable(torch.DoubleTensor(torch.randn(1, n_channel, feature_size, feature_size).double()), requires_grad=True), kernel_size)
     f_max_pooling = F.max_pool2d
     print gradcheck(f_max_pooling, input, eps=1e-3)
